Move Exception types to confidence.exceptions
