+++ confidence/__init__.py
@@ -8,37 +8,7 @@ import warnings
 
 import yaml
 
-
-class ConfigurationError(KeyError):
-    pass
-
-
-class MergeConflictError(ConfigurationError):
-    """
-    Error raised during loading configuration sources that conflict internally.
-    """
-    def __init__(self, *args, key):
-        super().__init__(*args)
-        self.conflict = key
-
-
-class NotConfiguredError(ConfigurationError):
-    """
-    Error raised when a requested configuration key is unavailable and no
-    default / fallback value is provided.
-    """
-    def __init__(self, *args, key):
-        super().__init__(*args)
-        self.key = key
-
-
-class ConfiguredReferenceError(ConfigurationError):
-    """
-    Error raised a referenced configuration key is unavailable.
-    """
-    def __init__(self, *args, key):
-        super().__init__(*args)
-        self.key = key
+from confidence.exceptions import ConfigurationError, ConfiguredReferenceError, MergeConflictError, NotConfiguredError
 
 
 class _Conflict(IntEnum):
